# Conflicts:
#	packages/rs-dpp/src/data_contract/data_contract_factory.rs
#	packages/wasm-dpp/src/errors/protocol_error.rs

+++ packages/wasm-dpp/src/errors/consensus/basic/identity/duplicated_identity_public_key_error.rs
@@ -21,7 +21,7 @@ impl DuplicatedIdentityPublicKeyErrorWasm {
         self.inner
             .duplicated_public_keys_ids()
             .iter()
-            .map(|id| JsValue::from(*id as u32))
+            .map(|id| JsValue::from(*id))
             .collect()
     }
