+++ packages/wasm-dpp/src/errors/consensus/state/identity/duplicated_identity_public_key_state_error.rs
@@ -14,7 +14,7 @@ impl DuplicatedIdentityPublicKeyStateErrorWasm {
         // TODO: key ids probably should be u32
         self.duplicated_public_keys_ids
             .iter()
-            .map(|id| JsValue::from(*id as u32))
+            .map(|id| JsValue::from(*id))
             .collect()
     }
